+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 
 # 2.0.0
 + **Breaking change:** Requires node 14 or above.
+
+# 3.0.0
++ **Breaking change:** The config format has been changed:
+    + Rule names are now camel cased.
+    + An `extends` property can be used to move common properties to other files.
+    + A config can now also be a filename, or an array of configs.
++ New command line flag added `--fix-duplicates`.